backend/pyproject.toml
[project]
name = "geminilangchainexercise1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "bcrypt<4.0.0",
    "langchain-tavily>=0.2.17",
    "mem0ai>=1.0.11",
    "langchain-pinecone>=0.2.13",
    "pinecone>=7.3.0",
    "qdrant-client>=1.17.1",
]
